Bug 1583967 - Fix pointer overflow errors in webrtc signal processing; r=ng
These have already been fixed upstream. The upstream comment was: "Negative
overflow is permitted here, because this is auto-regressive filters, and the
state for each batch run is stored in the "negative" positions of the output
vector."

Differential Revision: https://phabricator.services.mozilla.com/D48884

UltraBlame original commit: 1b3998d69371a2f9428f45551988580b5e00d348

+++ media/webrtc/trunk/webrtc/common_audio/signal_processing/downsample_fast.c
@@ -42,8 +42,8 @@ int WebRtcSpl_DownsampleFastC(const int16_t* data_in,
     out_s32 = 2048;  
 
     for (j = 0; j < coefficients_length; j++) {
-      rtc_MsanCheckInitialized(&data_in[i - j], sizeof(data_in[0]), 1);
-      out_s32 += coefficients[j] * data_in[i - j];  
+      rtc_MsanCheckInitialized(&data_in[(ptrdiff_t)i - (ptrdiff_t)j],sizeof(data_in[0]), 1);
+      out_s32 += coefficients[j] * data_in[(ptrdiff_t)i - (ptrdiff_t)j];  
     }
 
     out_s32 >>= 12;
